+++ preload.js
@@ -14,8 +14,8 @@ contextBridge.exposeInMainWorld('greenroom', {
   getMic: () => ipcRenderer.invoke('mic:get'),
   onMicState: (cb) => ipcRenderer.on('mic:state', (_e, v) => cb(v)),
 
-  // compact "Game Bar" view + hotkey status
-  onCompact: (cb) => ipcRenderer.on('ui:compact', (_e, v) => cb(v)),
+  // Game Bar overlay + hotkey status
+  onOverlay: (cb) => ipcRenderer.on('ui:overlay', (_e, v) => cb(v)),
   onHotkeyStatus: (cb) => ipcRenderer.on('hotkeys:status', (_e, v) => cb(v)),
 
   // settings